% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{addIPW}
\alias{addIPW}
\title{Add IPW(inverse probability weight) to a data.frame}
\usage{
addIPW(formula, data)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data.frame}
}
\description{
Add IPW(inverse probability weight) to a data.frame
}
\examples{
formula=treat~x1+x2
mydata=addIPW(treat~x1+x2,data=simData2)
}
