% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_mnps.R
\name{adjustedPlot_mnps}
\alias{adjustedPlot_mnps}
\title{Draw adjusted survival curve using weights}
\usage{
adjustedPlot_mnps(out, time = "", status = "", ...)
}
\arguments{
\item{out}{An object of class mnps}

\item{time}{Name of time variable}

\item{status}{Name of status variable}

\item{...}{further arguments to be passed survminer::ggsurvplot}
}
\value{
an object of class ggsurvplot
}
\description{
Draw adjusted survival curve using weights
}
\examples{
library(twang)
data(cancer,package="survival")
out=mnps(rx~nodes+differ+adhere+obstruct+surg+extent+node4,data=colon,verbose=FALSE,n.trees=3000)
adjustedPlot_mnps(out,time="time",status="status")
adjustedPlot_mnps(out,time="time",status="status",pval=TRUE,risk.table=TRUE)
}
