% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffect.R
\name{austinCafri}
\alias{austinCafri}
\title{Austin and Cafri's SE estimator}
\source{
Austin, Peter C., and Guy Cafri. 2020.
“Variance Estimation When Using Propensity-Score Matching with Replacement with Survival or Time-to-Event Outcomes.”
Statistics in Medicine 39 (11): 1623–40. \url{https://doi.org/10.1002/sim.8502}.
}
\usage{
austinCafri(formula, data)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data.frame which is the result of get_mactchs() function}
}
\description{
Austin and Cafri's SE estimator
}
\examples{
library(MatchIt)
library(survival)
out=matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData,
    link = 'linear.logit', caliper = .1,ratio = 3, replace = TRUE)
austinCafri(Surv(Y_S) ~ A,data=get_matches(out))
}
