% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_twangCont.R
\name{balance.plot}
\alias{balance.plot}
\title{Draw balance plot}
\usage{
balance.plot(x)
}
\arguments{
\item{x}{An object of class ps.cont}
}
\description{
Draw balance plot
}
\examples{
library(twangContinuous)
model=ps.cont( treat ~ x1 + x2,data=simData2,n.trees=5000)
balance.plot(model)
\dontrun{
data(dat)
model=ps.cont(tss_0~sfs8p_0+sati_0+sp_sm_0+recov_0+subsgrps_n+treat,data=dat,n.trees = 500)
balance.plot(model)
}
}
