% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{balanceCheck}
\alias{balanceCheck}
\title{Covariates balance check}
\usage{
balanceCheck(data, xvars = NULL, treatvar = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{xvars}{Names of covariates}

\item{treatvar}{Name of treatment var}
}
\description{
Covariates balance check
}
\examples{
data=addIPW(simData2,xvars=c("x1","x2"),treatvar="treat")
balanceCheck(data)
}
