% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_twangCont.R
\name{balancePlot}
\alias{balancePlot}
\title{Draw balance plot}
\usage{
balancePlot(x, show.point = TRUE, show.label = FALSE, show.legend = FALSE)
}
\arguments{
\item{x}{An object of class ps.cont}

\item{show.point}{logical Whether or not show point}

\item{show.label}{logical Whether or not show label}

\item{show.legend}{logical Whether or not show legend}
}
\description{
Draw balance plot
}
\examples{
\dontrun{
library(twangContinuous)
model=ps.cont( treat ~ x1 + x2,data=simData2,n.trees=5000)
balancePlot(model)
data(dat)
model=ps.cont(tss_0~sfs8p_0+sati_0+sp_sm_0+recov_0+subsgrps_n+treat,data=dat,n.trees = 500)
balancePlot(model)
}
}
