% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancePlotTwang.R
\name{balancePlotTwang}
\alias{balancePlotTwang}
\title{Draw balance Plot}
\usage{
balancePlotTwang(out, show.label = FALSE, show.legend = FALSE)
}
\arguments{
\item{out}{An object of class mnps}

\item{show.label}{logical Whether or not show label}

\item{show.legend}{logical Whether or not show legend}
}
\description{
Draw balance Plot
}
\examples{
library(twang)
data(AOD)
out=mnps(treat~illact+crimjust+subprob+subdep+white,data=AOD,n.trees=3000,verbose=FALSE,
stop.method=c("es.mean","ks.max"))
balancePlotTwang(out)
}
