% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{call2param}
\alias{call2param}
\title{Extract parameter from call}
\usage{
call2param(call)
}
\arguments{
\item{call}{An object of class call}
}
\value{
a list of parameters
}
\description{
Extract parameter from call
}
