% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{changeVarnames}
\alias{changeVarnames}
\title{Change names of factor or character column name}
\usage{
changeVarnames(data, xvars)
}
\arguments{
\item{data}{A data.frame}

\item{xvars}{string names of columns}
}
\description{
Change names of factor or character column name
}
\examples{
changeVarnames(iris,xvars=c("Septal.Length","Species"))
}
