% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM1.R
\name{compareLove.plot}
\alias{compareLove.plot}
\title{Drow love plot comparing to nearest match}
\usage{
compareLove.plot(out, stats = c("m"))
}
\arguments{
\item{out}{An object of a class matchit}

\item{stats}{character; which statistic(s) should be reported}
}
\description{
Drow love plot comparing to nearest match
}
\examples{
library(cobalt)
data(lalonde,package="MatchIt")
formula=treat ~ age + educ + race+ married +nodegree + re74 + re75
out=MatchIt::matchit(formula, data =lalonde, method= "full",link="probit")
compareLove.plot(out)
compareLove.plot(out,stats=c("m","ks"))
}
