% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM1.R
\name{descNames}
\alias{descNames}
\title{Describe names as natural language}
\usage{
descNames(x = LETTERS[1:5])
}
\arguments{
\item{x}{A charcter vector}
}
\description{
Describe names as natural language
}
\examples{
descNames("A")
descNames(c("A","B"))
descNames(LETTERS[1:5])
}
