% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffect.R
\name{estimateEffect}
\alias{estimateEffect}
\title{Estimate Effect after matching}
\usage{
estimateEffect(
  out,
  mode = "continuous",
  multiple = TRUE,
  dep,
  covarCentering = FALSE,
  withinSubclass = FALSE,
  digits = 2,
  sedigits = 2,
  pdigits = 4,
  se = TRUE,
  print = TRUE
)
}
\arguments{
\item{out}{An object of class "matchit"}

\item{mode}{One of c("continuous","binary","survival")}

\item{multiple}{logical Whether or not perform multiple regression}

\item{dep}{Name of dependent variable}

\item{covarCentering}{logical}

\item{withinSubclass}{logical}

\item{digits}{integer indicating the number of decimal places}

\item{sedigits}{digits for standard error}

\item{pdigits}{digits for p value}

\item{se}{logical If true, report se. If false report confidence interval}

\item{print}{logical}
}
\description{
Estimate Effect after matching
}
\examples{
library(MatchIt)
library(survival)
out <- matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData)
estimateEffect(out,dep=c("Y_C"),multiple=FALSE)
estimateEffect(out,dep=c("Y_C"))
estimateEffect(out,dep=c("Y_C","Y_B"))
estimateEffect(out,mode="binary",dep=c("Y_B"),multiple=FALSE)
estimateEffect(out,mode="binary",dep=c("Y_B"))
estimateEffect(out,mode="survival",dep=c("Y_S"),multiple=FALSE)
estimateEffect(out,mode="survival",dep=c("Y_S"))
\dontrun{
out=matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData,
   link = 'linear.logit', caliper = .1, ratio = 3, replace = TRUE)
estimateEffect(out,dep=c("Y_C"),multiple=FALSE)
estimateEffect(out,dep=c("Y_C"))
estimateEffect(out,mode="binary",dep=c("Y_B"),multiple=FALSE)
estimateEffect(out,mode="binary",dep=c("Y_B"))
estimateEffect(out,mode="survival",dep=c("Y_S"),multiple=FALSE)
estimateEffect(out,mode="survival",dep=c("Y_S"))
out=matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData,
method = "full", estimand = "ATE")
estimateEffect(out,dep=c("Y_C"),multiple=FALSE)
estimateEffect(out,dep=c("Y_C"),covarCentering=TRUE)
estimateEffect(out,mode="binary",dep=c("Y_B"),multiple=FALSE)
estimateEffect(out,mode="survival",dep=c("Y_S"),multiple=FALSE)
\dontrun{
out=matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData,
method = "subclass", estimand = "ATT",subclass = 8)
estimateEffect(out,dep=c("Y_C"),multiple=FALSE)
estimateEffect(out,dep=c("Y_C"),multiple=FALSE,withinSubclass=TRUE)
estimateEffect(out,mode="binary",dep=c("Y_B"),multiple=FALSE)
estimateEffect(out,mode="binary",dep=c("Y_B"))
estimateEffect(out,mode="survival",dep=c("Y_S"),multiple=FALSE)
out=matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData,
method = "full", estimand = "ATT")
estimateEffect(out,mode="binary",dep=c("Y_B"))
estimateEffect(out,mode="continuous",dep=c("Y_B"))  # for conditional effect
out=matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = swData,
method = "full", estimand = "ATE",s.weights="SW")
estimateEffect(out,dep=c("Y_C"))
x=matchit(treat ~ age + educ + race + married+nodegree + re74 + re75, data =lalonde,
   method="exact")
estimateEffect(x,dep=c("re78"))
}
}
