% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffect2.R
\name{estimateEffect2}
\alias{estimateEffect2}
\title{Estimate Effect after matching using the bootstrap}
\usage{
estimateEffect2(
  out,
  mode = "continuous",
  multiple = TRUE,
  dep,
  covarCentering = FALSE,
  withinSubclass = FALSE,
  seed = 1,
  digits = 2,
  sedigits = 2,
  pdigits = 4,
  se = TRUE,
  print = TRUE
)
}
\arguments{
\item{out}{An object of class "matchit"}

\item{mode}{One of c("continuous","binary","survival")}

\item{multiple}{logical Whether or not perform multiple regression}

\item{dep}{Name of dependent variable}

\item{covarCentering}{logical}

\item{withinSubclass}{logical}

\item{seed}{numeric}

\item{digits}{integer indicating the number of decimal places}

\item{sedigits}{digits for standard error}

\item{pdigits}{digits for p value}

\item{se}{logical If true, report se. If false report confidence interval}

\item{print}{logical}
}
\description{
Estimate Effect after matching using the bootstrap
}
\examples{
library(MatchIt)
library(survival)
out <- matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData)
\dontrun{
estimateEffect2(out,dep="Y_C",multiple=FALSE)
\dontrun{
estimateEffect2(out,mode="binary",dep="Y_B",multiple=FALSE)
estimateEffect2(out,mode="binary",dep="Y_B")
estimateEffect2(out,mode="survival",dep="Y_S",multiple=FALSE)
out <- matchit(A ~ X1 + X2 + X3 + X4 + X5 + X6 + X7 + X8 + X9, data = exData,
            link = 'linear.logit', caliper = .1, ratio = 3, replace = TRUE)
estimateEffect2(out,dep="Y_C",multiple=FALSE)
estimateEffect2(out,dep="Y_B",mode="binary",multiple=TRUE)
}
}
