% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{estimateEffectContinuous}
\alias{estimateEffectContinuous}
\title{estimate effect based on IPW(inverse probability weight)}
\usage{
estimateEffectContinuous(
  data,
  dep,
  xvars = NULL,
  treatvar = NULL,
  seed = 1234,
  probs = 0.1 * (1:9),
  num = 10000,
  weights = NULL
)
}
\arguments{
\item{data}{A data.frame as a result of addIPW()}

\item{dep}{Name of dependent variable}

\item{xvars}{Names of covariates}

\item{treatvar}{Name of treatment var}

\item{seed}{A single integer}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{num}{an integer specifying the number of simulations to compute}

\item{weights}{Name of weight variable}
}
\description{
estimate effect based on IPW(inverse probability weight)
}
\examples{
mydata=addIPW(treat~x1+x2,data=simData2)
estimateEffectContinuous(mydata,dep="y",weights="IPW")
\dontrun{
estimateEffectContinuous(mydata,dep="y")
data(lalonde,package="MatchIt")
mydata1=addIPW(age~educ+race+married, data=lalonde)
estimateEffectContinuous(mydata1,dep="re78",weights="IPW")
}
}
