% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_mnps.R
\name{estimateEffectTwang}
\alias{estimateEffectTwang}
\title{Estimate Treatment Effect}
\usage{
estimateEffectTwang(
  out,
  dep = "",
  time = "",
  status,
  stop.method = "es.mean",
  adjustCovar = FALSE,
  method = "GBM"
)
}
\arguments{
\item{out}{An object of class ps}

\item{dep}{Name of dependent variable}

\item{time}{Name of time variable}

\item{status}{Name of status variable}

\item{stop.method}{A method or methods of measuring and summarizing balance across pretreatment variables.}

\item{adjustCovar}{logical}

\item{method}{Method of PS estimate. One of c("GBM","lm","glm","chisq").}
}
\description{
Estimate Treatment Effect
}
\examples{
library(twang)
data(AOD)
\dontrun{
out=mnps(treat ~ illact + crimjust + subprob + subdep + white,data = AOD,
stop.method=c("es.mean","ks.mean"),verbose=FALSE,n.trees = 3000)
estimateEffectTwang(out,dep="suf12")
out1=mnps(treat ~ illact + crimjust + subprob + subdep + white,data = AOD,
estimand="ATT",treatATT="community",stop.method=c("es.mean","ks.mean"),verbose=FALSE,n.trees = 3000)
estimateEffectTwang(out1,dep="suf12")
data(cancer,package="survival")
fit=mnps(rx~nodes+differ+adhere+obstruct+surg+extent+node4,data=colon,verbose=FALSE,n.trees=3000)
result=estimateEffectTwang(fit,dep="status",time="time",status="status")
summary(result)
estimateEffectTwang(fit,dep="status")
}
}
