% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateEffectZelig.R
\name{estimateEffectZelig}
\alias{estimateEffectZelig}
\title{Estimate Effect after matching with simulation}
\usage{
estimateEffectZelig(
  out,
  mode = "continuous",
  multiple = TRUE,
  dep,
  seed = 1224,
  n_sim = 10000,
  reverse = FALSE
)
}
\arguments{
\item{out}{An object of class "matchit"}

\item{mode}{One of c("continuous","binary","survival")}

\item{multiple}{logical Whether or not perform multiple regression}

\item{dep}{Name of dependent variable}

\item{seed}{numeric seed number}

\item{n_sim}{numeric number of simuations}

\item{reverse}{logical}
}
\description{
Estimate Effect after matching with simulation
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",
caliper=0.15,ratio=2)
estimateEffectZelig(out,dep="y")
out1=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",estimand="ATC",
caliper=0.15,ratio=2)
estimateEffectZelig(out1,dep="y")
out2=matchit(formula=Rtreat~V1+V2+V3,data=simData,link="linear.logit",
caliper=0.15,ratio=2,replace=TRUE)
estimateEffectZelig(out2,dep="y",reverse=TRUE)
fullATT=matchit(formula=treat~V1+V2+V3,data=simData,
link="linear.logit",method="full")
estimateEffectZelig(fullATT,dep="y")
fullATC=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",
method="full",estimand="ATC")
estimateEffectZelig(fullATC,dep="y")
fullATC2=matchit(formula=Rtreat~V1+V2+V3,data=simData,link="linear.logit",
method="full")
estimateEffectZelig(fullATC2,dep="y",reverse=TRUE)
fullATE=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",
method="full",estimand="ATE")
estimateEffectZelig(fullATE,dep="y")
}
