% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectZeligSubclass.R
\name{estimateEffectZeligSubclass}
\alias{estimateEffectZeligSubclass}
\title{Estimate Effect after subclass matching with simulation}
\usage{
estimateEffectZeligSubclass(
  out,
  mode = "continuous",
  multiple = TRUE,
  dep,
  seed = 1224,
  n_sim = 10000,
  reverse = FALSE
)
}
\arguments{
\item{out}{An object of class "matchit"}

\item{mode}{One of c("continuous","binary","survival")}

\item{multiple}{logical Whether or not perform multiple regression}

\item{dep}{Name of dependent variable}

\item{seed}{numeric seed number}

\item{n_sim}{numeric number of simuations}

\item{reverse}{logical}
}
\description{
Estimate Effect after subclass matching with simulation
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,estimand="ATE",link="linear.logit",
method="subclass",subclass=8)
estimateEffectZeligSubclass(out,dep="y")
}
