% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exData.R
\docType{data}
\name{exData}
\alias{exData}
\title{Sample Data for test MatchIt package}
\format{
A data frame with 2000 rows and 13 variables
\describe{
 \item{A}{treatment variable, 0 or 1}
 \item{X1}{covariate 1}
 \item{X2}{covariate 2}
 \item{X3}{covariate 3}
 \item{X4}{covariate 4}
 \item{X5}{covariate 5}
 \item{X6}{covariate 6}
 \item{X7}{covariate 7}
 \item{X8}{covariate 8}
 \item{X9}{covariate 9}
 \item{Y_C}{continuous outcome}
 \item{Y_B}{binary outcome}
 \item{Y_S}{survival outcome}
}
}
\source{
\url{https://cran.r-project.org/web/packages/MatchIt/vignettes/estimating-effects.html}
}
\usage{
exData
}
\description{
A toy dataset containing treatment variable(A), numeric variables(X1-X9), continuous outcome(Y_C), binary outcome(Y_B) and survival outcome(Y_S).
}
\keyword{datasets}
