% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{factor2dummy}
\alias{factor2dummy}
\title{Make factor to dummary variable in a data.frame}
\usage{
factor2dummy(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Make factor to dummary variable in a data.frame
}
\examples{
iris2=factor2dummy(iris)
}
