% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptimalK.R
\name{findOptimalK}
\alias{findOptimalK}
\title{Find optimal number of groups for covariates for coarsened exact matching}
\usage{
findOptimalK(out, plot = FALSE, k = 2:10, mdthreshold = 0.1, nthreshold = 0.97)
}
\arguments{
\item{out}{An object of class matchit with coarsened exact matching}

\item{plot}{logical Whether or not draw plot}

\item{k}{The number of groups for continuous covariates}

\item{mdthreshold}{numeric thershold for absolute standardized mean difference}

\item{nthreshold}{numeric Threshold of ratio of matched treated group and treated group}
}
\value{
A list
\itemize{
\item df - A data.frame of standardized mean differences
\item cutpoints - A list of cutpoints
\item k Recommended number of groups for continuous covariates
\item p1 The first ggplot
\item p2 The 2nd ggplot
}
}
\description{
Find optimal number of groups for covariates for coarsened exact matching
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,method="cem")
estimateEffectZelig(out,dep="y")
result=findOptimalK(out)
out=matchit(formula=treat~V1+V2+V3,data=simData,method="cem",cutpoints=result$cutpoints)
estimateEffectZelig(out,dep="y")
out=matchit(treat ~ age + educ + race + married+nodegree + re74 + re75, data =lalonde,
method="cem",estimand="ATE")
findOptimalK(out)
}
