% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{formula2vars}
\alias{formula2vars}
\title{Extract variable names with formula}
\usage{
formula2vars(formula)
}
\arguments{
\item{formula}{An object of class formula}
}
\description{
Extract variable names with formula
}
\examples{
formula = treat ~ age + educ + race + married+nodegree + re74 + re75
formula2vars(formula)
}
