% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatchPair.R
\name{gammaRangeSearch}
\alias{gammaRangeSearch}
\title{Find significant gamma range}
\usage{
gammaRangeSearch(out, dep = "y", start = 1, threshold = 0.025, method = "w")
}
\arguments{
\item{out}{An object of class matchit}

\item{dep}{Character Name of dependent variable}

\item{start}{numeric start gamma value}

\item{threshold}{numeric p value threshold}

\item{method}{method to be passed to senmw()}
}
\description{
Find significant gamma range
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",caliper=0.15,ratio=2)
tail(gammaRangeSearch(out,dep="y"))
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",ratio=2,method="optimal")
tail(gammaRangeSearch(out,dep="y"))
out=matchit(treat~age+educ+race+married+nodegree+re74+re75,data=lalonde)
tail(gammaRangeSearch(out,dep="re78"))
}
