% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatchPairFull.R
\name{gammaRangeSearchFull}
\alias{gammaRangeSearchFull}
\title{Find significant gamma range for full matching}
\usage{
gammaRangeSearchFull(out, start = 1, threshold = 0.025)
}
\arguments{
\item{out}{An object of class matchit}

\item{start}{numeric start gamma value}

\item{threshold}{numeric p value threshold}
}
\description{
Find significant gamma range for full matching
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",method="full")
tail(gammaRangeSearchFull(out))
}
