% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatchPair.R
\name{generateMatchPair}
\alias{generateMatchPair}
\title{Generate data.frame for sensitivity test}
\usage{
generateMatchPair(out, dep = "y")
}
\arguments{
\item{out}{An object of class matchit}

\item{dep}{Name of dependent variable}
}
\description{
Generate data.frame for sensitivity test
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",caliper=0.15,ratio=2)
generateMatchPair(out,dep="y")
}
