% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatchPairFull.R
\name{generateMatchPairFull}
\alias{generateMatchPairFull}
\title{Generate list for sensitivity test}
\usage{
generateMatchPairFull(out, dep)
}
\arguments{
\item{out}{An object of class matchit}

\item{dep}{Name of dependent variable}
}
\description{
Generate list for sensitivity test
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",method="full")
generateMatchPairFull(out,dep="y")
}
