% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeightedValues.R
\name{getWeightedValues}
\alias{getWeightedValues}
\title{Get weighted value or p value}
\usage{
getWeightedValues(x, digits = 1)
}
\arguments{
\item{x}{An object of class matchit}

\item{digits}{numeric}
}
\description{
Get weighted value or p value
}
\examples{
require(MatchIt)
formula=treat ~ age + educ + race+married+nodegree + re74 + re75
x=matchit(formula, data =lalonde, method= "full",link="probit")
getWeightedValues(x)
}
