% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{ggPS}
\alias{ggPS}
\title{Make ggplot between covariates and propensity score}
\usage{
ggPS(x, xlabs = NULL, ncol = 2)
}
\arguments{
\item{x}{An object of class matchit}

\item{xlabs}{labels for group variable}

\item{ncol}{Numeric Number of facet column}
}
\description{
Make ggplot between covariates and propensity score
}
\examples{
require(MatchIt)
require(ggplot2)
formula=treat ~ age + race+educ + nodegree + re74 + re75
x=matchit(formula, data =lalonde, method= "nearest",ratio=1,caliper=0.25)
ggPS(x,xlabs=c("control","treated"))+theme(legend.position=c(0.8,0.15))
}
