% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{ggPSMSummary}
\alias{ggPSMSummary}
\title{Draw summary plot of propensity score matching}
\usage{
ggPSMSummary(x, show.table = TRUE, xpos = NULL, ypos = NULL)
}
\arguments{
\item{x}{An object of class matchit}

\item{show.table}{logical whether or not show table}

\item{xpos}{numeric Min x-axis position of table}

\item{ypos}{numeric Min y-axis position of table}
}
\description{
Draw summary plot of propensity score matching
}
\examples{
require(MatchIt)
formula=treat ~ age + race+educ + nodegree + re74 + re75
x=matchit(formula, data =lalonde, method= "nearest",ratio=1,caliper=0.25)
ggPSMSummary(x,show.table=FALSE)
ggPSMSummary(x)
}
