% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM1.R
\name{makeCompareBalTab}
\alias{makeCompareBalTab}
\title{Make balance table}
\usage{
makeCompareBalTab(out, print = TRUE)
}
\arguments{
\item{out}{An object of a class matchit}

\item{print}{logical}
}
\description{
Make balance table
}
\examples{
library(MatchIt)
library(cobalt)
formula=treat ~ age + educ + race+ married +nodegree + re74 + re75
out=matchit(formula, data =lalonde, method= "full",link="probit")
makeCompareBalTab(out)
}
