% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{makePPTList_IPW}
\alias{makePPTList_IPW}
\title{make pptList with Matching with IPW}
\usage{
makePPTList_IPW(dataName, xvars, treatvar, dep, seed = 1234)
}
\arguments{
\item{dataName}{Name of data}

\item{xvars}{Name of covariates}

\item{treatvar}{Name of continuous treatment variable}

\item{dep}{Name of dependent variable}

\item{seed}{Numeric}
}
\description{
make pptList with Matching with IPW
}
\examples{
result=makePPTList_IPW("simData2",xvars=c("x1","x2"),treatvar="treat",dep="y")
}
