% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{makePPTList_matchit}
\alias{makePPTList_matchit}
\title{make pptList with an object of class matchit}
\usage{
makePPTList_matchit(
  x,
  depvar = NULL,
  seed = 1234,
  compare = TRUE,
  report = TRUE,
  multiple = TRUE,
  depKind = "continuous",
  covarCentering = FALSE,
  withinSubclass = FALSE
)
}
\arguments{
\item{x}{An object of class matchit}

\item{depvar}{Variable name serves as dependent variables}

\item{seed}{Integer}

\item{compare}{logical}

\item{report}{logical}

\item{multiple}{logical}

\item{depKind}{character One of c("continuous","binary","survival")}

\item{covarCentering}{logical}

\item{withinSubclass}{logical}
}
\description{
make pptList with an object of class matchit
}
\examples{
require(MatchIt)
x=matchit(treat ~ age + educ + race + married+nodegree + re74 + re75, data =lalonde,
   method="subclass",subclass=4)
x=matchit(treat ~ age + educ + race + married+nodegree + re74 + re75, data =lalonde,
   method="exact")
x=matchit(treat ~ age + educ + race + married+nodegree + re74 + re75, data =lalonde,
   method="full",link='probit')
result=makePPTList_matchit(x)
result=makePPTList_matchit(x,depvar="re78")
}
