% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_twangCont.R
\name{makePPTList_twangCont}
\alias{makePPTList_twangCont}
\title{Make PPT list with continuous variable}
\usage{
makePPTList_twangCont(x, dep = "", seed = 1234)
}
\arguments{
\item{x}{string}

\item{dep}{character Name of dependent variable}

\item{seed}{numeric random seed}
}
\description{
Make PPT list with continuous variable
}
\examples{
library(twangContinuous)
library(survey)
x="ps.cont( treat ~ x1 + x2,data=simData2,n.trees=5000)"
result=makePPTList_twangCont(x,dep="y")
x="ps.cont(tss_0~sfs8p_0+sati_0+sp_sm_0+recov_0+subsgrps_n+treat,data=dat,n.trees = 500)"
result=makePPTList_twangCont(x,dep="sfs8p_3",seed=1234)
}
