% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{pformat}
\alias{pformat}
\title{Make p format}
\usage{
pformat(x, digits = 3)
}
\arguments{
\item{x}{string}

\item{digits}{numeric}
}
\description{
Make p format
}
\examples{
x=c("0.000","0.123","","0.34")
pformat(x)
}
