% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{PSMTable}
\alias{PSMTable}
\title{Make flextable summarizing propensity score matching}
\usage{
PSMTable(x, digitsstd = 3, grouplabel = NULL)
}
\arguments{
\item{x}{An object of class matchit}

\item{digitsstd}{integer indicating the number of decimal places}

\item{grouplabel}{Optional group label}
}
\description{
Make flextable summarizing propensity score matching
}
\examples{
require(MatchIt)
formula=treat ~ age + educ + race+married+nodegree + re74 + re75
x=matchit(formula, data =lalonde, method= "full")
PSMTable(x,grouplabel=c("Control","Treated"))
x=matchit(formula, data =lalonde, method= "nearest")
PSMTable(x,grouplabel=c("Control","Treated"))
}
