% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM1.R
\name{reportEffect}
\alias{reportEffect}
\title{Report treatment Effect}
\usage{
reportEffect(
  out,
  yvar = "treat",
  depvar = "re78",
  digits = 2,
  sedigits = 2,
  pdigits = 4,
  se = TRUE
)
}
\arguments{
\item{out}{An object of class lm}

\item{yvar}{The name of group variable}

\item{depvar}{The name of dependent variable}

\item{digits}{integer indicating the number of decimal places}

\item{sedigits}{digits for standard error}

\item{pdigits}{digits for p value}

\item{se}{logical If true, report se. If false report confidence interval}
}
\description{
Report treatment Effect
}
\examples{
library(MatchIt)
library(sandwich)
formula= treat ~ age + educ + race+ married +nodegree + re74 + re75
out=matchit(formula, data =lalonde, method= "full",link="probit")
formula1= re78~treat +age + educ + race+ married +nodegree + re74 + re75
out1=lm(formula1,data=match.data(out),weights=weights)
reportEffect(out1,depvar="re78")
reportEffect(out1,depvar="re78",se=FALSE)
}
