% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM1.R
\name{reportPSM}
\alias{reportPSM}
\title{Report propensity sore matching}
\usage{
reportPSM(out, depvar = "", compare = NULL)
}
\arguments{
\item{out}{An object of class matchit}

\item{depvar}{Character name of dependent variable}

\item{compare}{logical Whether or not compare to nearest model}
}
\description{
Report propensity sore matching
}
\examples{
library(MatchIt)
formula=treat ~ age + educ + race+ married +nodegree + re74 + re75
out=matchit(formula, data =lalonde, method= "nearest")
reportPSM(out)
out=matchit(formula, data = lalonde, method= "full",distance="glm",link="probit")
reportPSM(out,depvar="re78")
reportPSM(out,depvar="re78",compare=FALSE)
}
