% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM1.R
\name{reportSMD}
\alias{reportSMD}
\title{Report adequacy of balance}
\usage{
reportSMD(out)
}
\arguments{
\item{out}{An object of class matchit}
}
\description{
Report adequacy of balance
}
\examples{
library(MatchIt)
formula=treat ~ age + educ + race+ married +nodegree + re74 + re75
out=matchit(formula, data =lalonde, method= "nearest")
reportSMD(out)
out1=matchit(formula, data =lalonde, method= "full",link="probit")
reportSMD(out1)
}
