% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData2.R
\docType{data}
\name{simData2}
\alias{simData2}
\title{simulation data by Young Min Baek}
\format{
A data frame with 1000 rows and 6 variables
\describe{
 \item{x1}{covariate 1}
 \item{x2}{covariate 2}
 \item{treat}{continuouis treatment variable}
 \item{y}{continuous outcome}
}
}
\source{
\url{https://sites.google.com/site/ymbaek}
}
\usage{
simData2
}
\description{
A toy dataset containing numeric variables(x1 and x2), a continuous variable(treat) and a continuous outcome(y).
}
\keyword{datasets}
