% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{standardize}
\alias{standardize}
\title{Standardize}
\usage{
standardize(x, max.ylev = 2)
}
\arguments{
\item{x}{A vector}

\item{max.ylev}{Numeric maximum length of unique values to be treated as a dummy variable}
}
\description{
Standardize
}
