% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{zelig2est}
\alias{zelig2est}
\title{Estimate dose-response function}
\usage{
zelig2est(out, probs = 0.1 * (1:9), num = 10000)
}
\arguments{
\item{out}{A object as a result of zelig()}

\item{probs}{numeric vector of probabilities with values in [0,1]}

\item{num}{an integer specifying the number of simulations to compute}
}
\description{
Estimate dose-response function
}
\examples{
library(Zelig)
mydata=addIPW(simData2,xvars=c("x1","x2"),treatvar="treat")
out=zelig(y~treat+x1+x2,data=mydata,model="ls",weights="IPW",cite=FALSE)
zelig2est(out)
}
