% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffectContinuous.R
\name{addIPW}
\alias{addIPW}
\title{Add IPW(inverse probability weight) to a data.frame}
\usage{
addIPW(data, xvars, treatvar)
}
\arguments{
\item{data}{A data.frame}

\item{xvars}{Names of covariates}

\item{treatvar}{Name of continuous treatment variable}
}
\description{
Add IPW(inverse probability weight) to a data.frame
}
\examples{
mydata=addIPW(simData2,xvars=c("x1","x2"),treatvar="treat")
}
