% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatchPair.R
\name{gammaRangeSearch}
\alias{gammaRangeSearch}
\title{Find significant gamma range}
\usage{
gammaRangeSearch(data, start = 1, threshold = 0.025, method = "w")
}
\arguments{
\item{data}{A data.frame as a result of generateMatchPair()}

\item{start}{numeric start gamma value}

\item{threshold}{numeric p value threshold}

\item{method}{method to be passed to senmw()}
}
\description{
Find significant gamma range
}
\examples{
library(MatchIt)
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",caliper=0.15,ratio=2)
df=generateMatchPair(out,dep="y")
tail(gammaRangeSearch(df))
out=matchit(formula=treat~V1+V2+V3,data=simData,link="linear.logit",ratio=2,method="optimal")
df=generateMatchPair(out,dep="y")
tail(gammaRangeSearch(df))
}
