% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeightedValues.R
\name{getWeightedValues}
\alias{getWeightedValues}
\title{Get weighted value or p value}
\usage{
getWeightedValues(x, xvar, treat = 1, digits = 1)
}
\arguments{
\item{x}{An object of class matchit}

\item{xvar}{character colume name}

\item{treat}{numeric  If 0, control. If 1, treat. If 2 p value.}

\item{digits}{numeric}
}
\description{
Get weighted value or p value
}
