% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTList_matchit.R
\name{makePPTList_matchit}
\alias{makePPTList_matchit}
\title{make pptList with an object of class matchit}
\usage{
makePPTList_matchit(x, depvar = NULL, compare = TRUE, report = TRUE)
}
\arguments{
\item{x}{An object of class matchit}

\item{depvar}{Variable name serves as dependent variables}

\item{compare}{logical}

\item{report}{logical}
}
\description{
make pptList with an object of class matchit
}
\examples{
require(MatchIt)
x="matchit(treat ~ age + educ + race + married+nodegree + re74 + re75, data =lalonde,
 method= 'nearest',ratio=1,link='probit')"
result=makePPTList_matchit(x)
result=makePPTList_matchit(x,depvar="re78")
}
